+++ classes/Notifiers/Expiration.class.php
@@ -63,7 +63,7 @@ class Expiration extends \Membership\BaseNotifier
             } else {
                 // Get the members based on notification counter and expiration
                 $qb->where('m.mem_notified > 0')
-                   ->andWhere('m.mem_expires < DATE_ADD(now(), INTERVAL (m.mem_notified * :interval) DAY')
+                   ->andWhere('m.mem_expires < DATE_ADD(now(), INTERVAL ((m.mem_notified -1) * :interval) DAY')
                    ->andWhere('m.mem_status IN (:stat)');
                    ->setParameter('interval', $interval, Database::INTEGER)
                    ->setParameter('stat', $stats);